backend/requirements.txt
fastapi[standard]==0.110.0
uvicorn==0.27.1
pydantic>=2.0.0
pydantic-settings
supabase
python-dotenv==1.0.1
google-auth==2.28.1
PyJWT>=2.10.0
tavily-python==0.3.3
google-generativeai>=0.5.0
reportlab==4.1.0
httpx
tortoise-orm==0.20.0
asyncpg==0.29.0
python-dateutil==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
websockets>=13.1
authlib==1.3.0
itsdangerous==2.1.2
